Canada Unemployment Rate Statistics--Tableau
The Canada Unemployment Rate Statistics Dashboard in Tableau provides a comprehensive overview of the unemployment rate trends and key statistics in Canada. It offers interactive visualizations and filters to allow users to explore the data in a user-friendly manner. the Kaggle dataset "Unemployment in Canada, by Province (1976-Present)" was used