KNN ML project with classified data
We are given a classified data set from a company (Very similar to a lot of data science interview projects, so the companies can protect their information). They've hidden the feature column names but have given the data and the target classes. 

We try to use KNN to create a model that directly predicts a class for a new data point based off of the features